feat: update styles to match bem #36
