+++ src/index.css
@@ -25,50 +25,3 @@ button {
   background-color: #F2CC39;
   cursor: pointer;
 }
-
-.main {
-  padding: 48px 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.title {
-  margin-bottom: 0;
-  font-size: 22px;
-  text-align: center;
-}
-
-.docs {
-  margin-bottom: 22px;
-  text-align: center;
-}
-
-.docs a {
-  font-size: 13px;
-  color: #506AD4;
-}
-
-.description {
-  margin-top: 6px;
-}
-
-.section {
-  min-width: 946px;
-}
-
-.section-title {
-  margin-bottom: 10px;
-}
-
-.code-container {
-  padding: 8px;
-  margin-top: 6px;
-  background-color: white;
-  border: 1px solid #C2B8AD;
-  border-radius: 6px;
-}
-
-.show-code {
-  margin-top: 16px;
-}
