+++ src/show-code/show-code.css
@@ -1,0 +1,11 @@
+.show-code {
+  margin-top: 16px;
+}
+
+.show-code__code-container {
+  padding: 8px;
+  margin-top: 6px;
+  background-color: white;
+  border: 1px solid #C2B8AD;
+  border-radius: 6px;
+}
