+++ src/show-code/show-code.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react"
+import "./show-code.css"
 
 interface Props {
   code: string
@@ -13,7 +14,7 @@ export const ShowCode = ({ code }: Props) => {
     <div className="show-code">
       <button onClick={toggle}>{shouldShow ? "Hide code" : "Show code"}</button>
 
-      {shouldShow && <div className="code-container"><code className="code">{code}</code></div>}
+      {shouldShow && <div className="show_code__code-container"><code>{code}</code></div>}
     </div>
   )
 }
